requirements.txt
aiodns==3.2.0
aiohttp==3.9.4
aiosignal==1.3.1
anyio==4.3.0
appnope==0.1.4
asttokens==2.4.1
attrs==23.2.0
ccxt==4.2.96
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
comm==0.2.2
cryptography==42.0.5
debugpy==1.8.1
decorator==5.1.1
executing==2.0.1
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
ipykernel==6.29.4
ipython==8.23.0
jedi==0.19.1
jupyter_client==8.6.1
jupyter_core==5.7.2
matplotlib-inline==0.1.6
multidict==6.0.5
nest-asyncio==1.6.0
packaging==24.0
parso==0.8.4
pexpect==4.9.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pycares==4.4.0
pycparser==2.22
Pygments==2.17.2
python-dateutil==2.9.0.post0
python-telegram-bot==21.1
pyzmq==25.1.2
requests==2.31.0
setuptools==69.2.0
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
tornado==6.4
traitlets==5.14.2
typing_extensions==4.11.0
urllib3==2.2.1
wcwidth==0.2.13
yarl==1.9.4
